requirements.txt
fastapi~=0.115.12
psutil~=7.0.0
uvicorn
motor
python-dotenv